+++ static/resources/user/profileHeaderImageUpload.js
@@ -1,7 +1,7 @@
 /**
  * Handle Image Upload
  */
-$(document).on('ready pjax:success', function () {
+$(document).on('humhub:ready', function () {
     'use strict';
 
     $('.fileupload').each(function () {
@@ -91,7 +91,7 @@ $(document).on('ready pjax:success', function () {
     });
 
 
-})
+});
 
 
 /**
@@ -121,7 +121,7 @@ function resetProfileImage(json) {
     $('.image-upload-buttons').hide();
 }
 
-$(document).on('ready pjax:success', function () {
+$(document).on('humhub:ready', function () {
 
     // override standard drag and drop behavior
     $(document).off('drop.humhub dragover.humhub').on('drop.humhub dragover.humhub', function (e) {